CAS integration with the paywall
it's a bit weird though. AFAICT, the Drupal end is not doing a redirect
that it's supposed to do.

* you click the CAS LOGIN button on the Django app
* it takes you to:

    https://teachrecovery.cumc.columbia.edu/cas/login?service=https://teachrecovery.ccnmtl.columbia.edu/accounts/caslogin/?next=/

* which immediately redirects you to:

    https://teachrecovery.cumc.columbia.edu/user?destination=cas/login

* when you enter a valid username/password there, it just deposits you
  at:

    https://teachrecovery.cumc.columbia.edu/

* if you go back to the django app and hit the CAS LOGIN button now, it
  properly redirects back to our cas endpoint with a valid ticketid and
  Django can log you in.

So, the issue is that somewhere on the Drupal side, instead of
redirecting to the specified 'service' URL with a ticketid immediately
after logging in, it instead just redirects to what I assume is some
kind of Drupal default login.

I dug up the PHP code for the Drupal CAS provider and the relevent bit
looks like:

```
function cas_server_login() {
  // Set login cookie so that we know we're in the process of logging in
  global $user;
  $output='';
  $service = isset($_REQUEST['service']) ? $_REQUEST['service'] : '';
  $gateway = isset($_REQUEST['gateway']);
  if ($user->uid) {
    if ($service) {
      $_COOKIE[CAS_LOGIN_COOKIE] = $service;
    }
    $output=t('You have successfully logged into CAS');
    cas_server_service_return();
  }
  else {
    if ($gateway && $service) {
      drupal_goto($service);
    }
    else {
      // Redirect to user login
      if ($service) {
        setcookie(CAS_LOGIN_COOKIE, $service);
      }
      $output .= l(t('Login'), 'user', array('query' => array('destination' => 'cas/login')));
      drupal_goto('user', array('query' => array('destination' => 'cas/login')));
    }
  }
  return $output;
}
```

which basically looks OK.

If you're not already logged in, it stashes the service URL into a
cookie (which I've verified) and then redirects to the local Drupal
login with 'cas/login' as the destination. The next time the browser
ends up back there, it pulls the service URL back out of the cookie and
redirects to it with a ticket.

It just appears that something else somewhere else in the Drupal setup
is interfering and sends it back to the site index instead of the
specified destination after login.

Anyway, at this point, I'm fairly certain that things are correct
on *our* end, so this should go out so we can start pressing to get the
Drupal end fixed up.